implement CycleCommandExecutor for managing player commands with permission checks and instruction execution
